requirements.txt
deepagents
deepagents-cli
pyyaml
